Move `eCarDrivingStyle` to an enum header
