+++ source/game_sa/Enums/eCarDrivingStyle.h
@@ -1,0 +1,10 @@
+#pragma once
+
+enum eCarDrivingStyle : int8
+{
+    DRIVINGSTYLE_STOP_FOR_CARS,
+    DRIVINGSTYLE_SLOW_DOWN_FOR_CARS,
+    DRIVINGSTYLE_AVOID_CARS,
+    DRIVINGSTYLE_PLOUGH_THROUGH,
+    DRIVINGSTYLE_STOP_FOR_CARS_IGNORE_LIGHTS
+};
